Add files via upload
    +-------+-------+-------+
    |             |            |           |
    |   1        |   2       |   3      |
    |             |            |           |
    +-------+-------+-------+
    |             |            |           |
    |   4        |   x       |   6      |
    |             |            |           |
    +-------+-------+-------+
    |             |            |           |
    |   7        |   8       |   9      |
    |             |            |           |
    +-------+-------+-------+

A game of 'Tic Tac Toe' between a player (mark 'o') and computer (mark 'x'). 
The player makes their move by entering an integer corresponding to the cell they want to place their mark in.
First move is always by the computer and is placed in the middle of the board (cell no. 5)
The first party to place three of the marks in a row (diagonally, vertically or horizontally) is the winner.

A game written in purely procedural Python. To be played on command line (Windows).